% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_hw_data}
\alias{load_hw_data}
\title{Easily filter unemployment data for Luxembourg}
\usage{
load_hw_data(file)
}
\arguments{
\item{file}{A data frame containing unemployment data for Luxembourg.}
}
\value{
A data frame
}
\description{
Easily filter unemployment data for Luxembourg
}
\details{
This function allows the user to load a sex-weight-height csv file. The file
is then transformed into a tibble.
}
