% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_height.R
\name{plot_h_density}
\alias{plot_h_density}
\title{Easily filter unemployment data for Luxembourg}
\usage{
plot_h_density(data)
}
\arguments{
\item{data}{A tibble containing sex-height-weight data.}
}
\value{
A ggplot
}
\description{
Easily filter unemployment data for Luxembourg
}
\details{
This function allows the user to create density plot of the height depending on sex.
Moreover, it takes displays the mean height of each sex as a dashed line.
}
